+++ src/mine/iterator.cpp
@@ -97,10 +97,10 @@ bool Iterator::inc(Operation& op) {
       return true;
 
     case Operation::Type::BIN:
-      op.type = Operation::Type::CMP;
+      op.type = Operation::Type::EQU;
       return true;
 
-    case Operation::Type::CMP:
+    case Operation::Type::EQU:
       op.type = Operation::Type::LPB;
       return true;
 
@@ -140,7 +140,7 @@ bool Iterator::shouldSkip(const Operation& op) {
        op.type == Operation::Type::TRN || op.type == Operation::Type::MUL ||
        op.type == Operation::Type::DIV || op.type == Operation::Type::DIF ||
        op.type == Operation::Type::MOD || op.type == Operation::Type::GCD ||
-       op.type == Operation::Type::BIN || op.type == Operation::Type::CMP)) {
+       op.type == Operation::Type::BIN || op.type == Operation::Type::EQU)) {
     return true;
   }
   if (op.source == CONSTANT_ZERO &&